feat: joinguild and leaveguild functionality added

+++ Orange.Api/Models/Guild.cs
@@ -7,5 +7,5 @@ public class Guild
 {
     public int Id { get; set; }
     
-    public uint GuildId { get; set; }
+    public ulong GuildId { get; set; }
 }
